Move Kafka consumer/producer into a clustered submodule
This is preparation for new communication alternatives like direct node-to-node communication
