+++ cache/clustered/clustered/src/main/java/org/eclipse/datagrid/cache/clustered/types/ClusteredCacheMessageReceiver.java
@@ -14,146 +14,9 @@ package org.eclipse.datagrid.cache.clustered.types;
  * #L%
  */
 
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.common.errors.InterruptException;
-import org.apache.kafka.common.serialization.ByteArrayDeserializer;
-import org.apache.kafka.common.serialization.VoidDeserializer;
-import org.eclipse.serializer.Serializer;
 import org.eclipse.serializer.typing.Disposable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.nio.charset.StandardCharsets;
-import java.time.Duration;
-import java.util.Collections;
-import java.util.Properties;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-public class ClusteredCacheMessageReceiver implements Disposable
+public interface ClusteredCacheMessageReceiver extends Disposable
 {
-    private static final Logger logger = LoggerFactory.getLogger(ClusteredCacheMessageReceiver.class);
-    private final Properties kafkaProperties;
-    private final String topicName;
-    private final String clientId;
-    private final ClusteredCacheMessageAcceptor messageAcceptor;
-    private final Serializer<byte[]> serializer;
-
-    private boolean started = false;
-    private final AtomicBoolean active = new AtomicBoolean(false);
-    private Thread thread;
-
-    public ClusteredCacheMessageReceiver(
-        final Properties kafkaProperties,
-        final String topicName,
-        final String clientId,
-        final ClusteredCacheMessageAcceptor messageAcceptor,
-        final Serializer<byte[]> serializer
-    )
-    {
-        this.kafkaProperties = kafkaProperties;
-        this.topicName = topicName;
-        this.clientId = clientId;
-        this.messageAcceptor = messageAcceptor;
-        this.serializer = serializer;
-    }
-
-    public void start() throws IllegalStateException
-    {
-        if (this.started)
-        {
-            throw new IllegalStateException("Receiver is already running");
-        }
-
-        this.thread = new Thread(this::run);
-        this.thread.start();
-
-        this.started = true;
-        this.active.set(true);
-    }
-
-    private void run()
-    {
-        final Properties properties = new Properties();
-        properties.putAll(this.kafkaProperties);
-        properties.put(ConsumerConfig.GROUP_ID_CONFIG, this.clientId);
-        properties.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "latest");
-        properties.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, VoidDeserializer.class.getName());
-        properties.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, ByteArrayDeserializer.class.getName());
-
-        try (final var consumer = new KafkaConsumer<String, byte[]>(properties))
-        {
-            consumer.subscribe(Collections.singleton(this.topicName));
-            while (this.active.get())
-            {
-                final ConsumerRecords<String, byte[]> records;
-
-                try
-                {
-                    records = consumer.poll(Duration.ofMillis(Long.MAX_VALUE));
-                }
-                catch (final InterruptException e)
-                {
-                    // clear interrupt flag so Kafka can properly close the consumer
-                    final var ignored = Thread.interrupted();
-                    continue;
-                }
-
-                this.consume(records);
-
-                try
-                {
-                    consumer.commitSync();
-                }
-                catch (final InterruptException e)
-                {
-                    // clear interrupt flag so Kafka can properly close the consumer
-                    final var ignored = Thread.interrupted();
-                }
-            }
-        }
-    }
-
-    private void consume(final ConsumerRecords<String, byte[]> records)
-    {
-        for (final var record : records)
-        {
-            final String senderId = new String(
-                record.headers().lastHeader("Sender-Id").value(),
-                StandardCharsets.UTF_8
-            );
-            if (senderId.equals(this.clientId))
-            {
-                continue;
-            }
-            final ClusteredCacheMessage message = this.serializer.deserialize(record.value());
-            this.messageAcceptor.accept(message);
-        }
-    }
-
-    @Override
-    public void dispose()
-    {
-        this.active.set(false);
-
-        if (this.thread != null)
-        {
-            this.thread.interrupt();
-
-            try
-            {
-                this.thread.join(1_000L);
-            }
-            catch (final InterruptedException e)
-            {
-                // ignored
-            }
-
-            if (this.thread.isAlive())
-            {
-                logger.warn("Failed to wait for receiver thread to finish.");
-            }
-        }
-    }
+    void start();
 }
